gst-plugins-bad1.0 Debian release 1.5.2-1

Conflicts:
	debian/build-deps
	debian/build-deps.in
	debian/changelog
	debian/control
	debian/control.in
	debian/gstreamer-plugins-bad.install
	debian/libgstreamer-plugins-bad-dev.install
	debian/libgstreamer-plugins-bad.install
	debian/rules

+++ gst-libs/gst/mpegts/gst-dvb-section.c
@@ -13,7 +13,7 @@
  *   Zaheer Abbas Merali <zaheerabbas at merali dot org>
  *   Edward Hervey <edward@collabora.com>
  *
- * This library is free softwagre; you can redistribute it and/or
+ * This library is free software; you can redistribute it and/or
  * modify it under the terms of the GNU Library General Public
  * License as published by the Free Software Foundation; either
  * version 2 of the License, or (at your option) any later version.
@@ -360,7 +360,7 @@ _parse_bat (GstMpegtsSection * section)
     g_ptr_array_add (bat->streams, stream);
 
     if (transport_stream_loop_length < 6) {
-      /* each entry must be at least 6 bytes (+ 4bytes CRC) */
+      /* each entry must be at least 6 bytes (+ 4 bytes CRC) */
       GST_WARNING ("PID %d invalid BAT entry size %d",
           section->pid, transport_stream_loop_length);
       goto error;
@@ -383,7 +383,7 @@ _parse_bat (GstMpegtsSection * section)
       GST_WARNING
           ("PID %d invalid BAT entry %d descriptors loop length %d (only have %"
           G_GSIZE_FORMAT ")", section->pid, section->subtable_extension,
-          descriptors_loop_length, end - 4 - data);
+          descriptors_loop_length, (gsize) (end - 4 - data));
       goto error;
     }
     stream->descriptors =
@@ -541,7 +541,7 @@ _parse_nit (GstMpegtsSection * section)
     g_ptr_array_add (nit->streams, stream);
 
     if (transport_stream_loop_length < 6) {
-      /* each entry must be at least 6 bytes (+ 4bytes CRC) */
+      /* each entry must be at least 6 bytes (+ 4 bytes CRC) */
       GST_WARNING ("PID %d invalid NIT entry size %d",
           section->pid, transport_stream_loop_length);
       goto error;
@@ -564,7 +564,7 @@ _parse_nit (GstMpegtsSection * section)
       GST_WARNING
           ("PID %d invalid NIT entry %d descriptors loop length %d (only have %"
           G_GSIZE_FORMAT ")", section->pid, section->subtable_extension,
-          descriptors_loop_length, end - 4 - data);
+          descriptors_loop_length, (gsize) (end - 4 - data));
       goto error;
     }
     stream->descriptors =
